// src/app/models/claims.enum.ts
export enum Claims {
    SidTokenKey = 'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/sid',
    NameTokenKey = 'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name',
    EmailTokenKey = 'http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress',
    RoleTokenKey = 'http://schemas.microsoft.com/ws/2008/06/identity/claims/role',
    UserDataTokenKey = 'http://schemas.microsoft.com/ws/2008/06/identity/claims/userdata'
  }